# Conflicts:
#	src/main/java/com/sparta/trellocloneproject/controller/board/BoardController.java
#	src/main/java/com/sparta/trellocloneproject/entity/BoardMember.java
#	src/main/java/com/sparta/trellocloneproject/repository/BoardMemberRepository.java
#	src/main/java/com/sparta/trellocloneproject/service/board/BoardMemberService.java
#	src/main/java/com/sparta/trellocloneproject/service/board/BoardService.java

+++ src/main/java/com/sparta/trellocloneproject/controller/board/BoardController.java
@@ -6,6 +6,10 @@ import com.sparta.trellocloneproject.entity.Board;
 import com.sparta.trellocloneproject.security.UserDetailsImpl;
 import com.sparta.trellocloneproject.service.board.BoardMemberService;
 import com.sparta.trellocloneproject.service.board.BoardService;
+import com.sparta.trellocloneproject.dto.board.requestDto.BoardRequestDto;
+import com.sparta.trellocloneproject.dto.board.requestDto.BoardUpdateColorDto;
+import com.sparta.trellocloneproject.dto.board.requestDto.BoardUpdateDescriptionDto;
+import com.sparta.trellocloneproject.dto.board.requestDto.BoardUpdateTitleDto;
 import com.sparta.trellocloneproject.dto.board.responseDto.BoardResponseDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -55,11 +59,12 @@ public class BoardController {
             @PathVariable Long boardId,
             @RequestBody BoardUpdateTitleDto requestTitle,
             @AuthenticationPrincipal UserDetailsImpl userDetails) {
-
+        if(boardMemberService.isUserMember(userDetails,boardId)){
         Board board = boardService.updateBoardTitle(boardId, requestTitle, userDetails.getUser());
         BoardResponseDto boardResponseDto = new BoardResponseDto(board.getTitle());
 
-        return ResponseEntity.ok(boardResponseDto);
+        return ResponseEntity.ok(boardResponseDto);}
+        return null;
     }
 
     @PutMapping("/{boardId}/color")
@@ -67,11 +72,13 @@ public class BoardController {
             @PathVariable Long boardId,
             @RequestBody BoardUpdateColorDto requestColor,
             @AuthenticationPrincipal UserDetailsImpl userDetails) {
-
+        if(boardMemberService.isUserMember(userDetails,boardId)){
         Board board = boardService.updateBoardColor(boardId, requestColor, userDetails.getUser());
         BoardResponseDto boardResponseDto = new BoardResponseDto(board.getTitle());
 
         return ResponseEntity.ok(boardResponseDto);
+        }
+        return null;
     }
 
     @PutMapping("/{boardId}/description")
@@ -79,27 +86,35 @@ public class BoardController {
             @PathVariable Long boardId,
             @RequestBody BoardUpdateDescriptionDto requestDescription,
             @AuthenticationPrincipal UserDetailsImpl userDetails) {
-
+        if(boardMemberService.isUserMember(userDetails,boardId)){
         Board board = boardService.updateBoardDescription(boardId, requestDescription, userDetails.getUser());
         BoardResponseDto boardResponseDto = new BoardResponseDto(board.getTitle());
 
         return ResponseEntity.ok(boardResponseDto);
+        }
+        return null;
     }
 
     @DeleteMapping("/{boardId}")
     public ResponseEntity<String> deleteBoard(
             @PathVariable Long boardId,
             @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        if(boardMemberService.isUserMember(userDetails,boardId)){
         boardService.deleteBoard(boardId, userDetails.getUser());
 
         return ResponseEntity.ok("보드 삭제 완료");
+        }
+        return null;
     }
 
     @PostMapping("/{boardId}/members")
-    public String addBoardMember(@RequestBody BoardMemberRequestDto requestDto, @PathVariable Long boardId, @AuthenticationPrincipal UserDetailsImpl userDetails) {
+    public String addmember(@RequestBody BoardMemberRequestDto requestDto, @PathVariable Long boardId, @AuthenticationPrincipal UserDetailsImpl userDetails){
+        if(boardMemberService.isUserMember(userDetails,boardId)){
         Long UId = requestDto.getUserId();
         boardMemberService.addBoardMember(UId, boardId);
         return "ok";
+        }
+        return null;
     }
 
     @DeleteMapping("/{boardId}/member")
@@ -111,5 +126,4 @@ public class BoardController {
 
         return ResponseEntity.ok("보드 멤버 삭제 완료");
     }
-
 }
